ai-legal-engine/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Gemini AI (FREE!)
google-generativeai==0.3.1

# Document Processing (Pure Python - NO C++ needed!)
pypdf==3.17.1
python-docx==1.1.0
python-multipart==0.0.6

# Text Processing (Pure Python)
nltk==3.8.1

# Utilities
httpx==0.25.2
tenacity==8.2.3

# We removed: aiohttp (has C dependencies), orjson (needs Rust), textstat (not critical), chromadb (not needed yet)
